@@ -28,69 +28,69 @@
     "i18n": "i18next --config i18next-parser.config.js"
   },
   "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@ant-design/pro-card": "^1.19.5",
-    "@ant-design/pro-form": "^1.57.0",
-    "@ant-design/pro-table": "^2.66.4",
-    "@ant-design/pro-utils": "^1.39.2",
-    "@monaco-editor/react": "^4.4.4",
-    "ahooks": "^3.1.13",
-    "antd": "^4.20.5",
-    "antd-img-crop": "^4.1.0",
-    "axios": "^0.26.1",
-    "copy-to-clipboard": "^3.3.1",
+    "@ant-design/icons": "^5.0.1",
+    "@ant-design/pro-card": "^2.1.10",
+    "@ant-design/pro-form": "^2.5.4",
+    "@ant-design/pro-table": "^3.2.12",
+    "@ant-design/pro-utils": "^2.5.4",
+    "@monaco-editor/react": "^4.4.6",
+    "ahooks": "^3.7.4",
+    "antd": "^5.1.6",
+    "antd-img-crop": "^4.6.0",
+    "axios": "^1.2.5",
+    "copy-to-clipboard": "^3.3.3",
     "crypto-js": "^4.1.1",
-    "i18next": "^21.6.14",
-    "i18next-http-backend": "^1.4.0",
+    "i18next": "^22.4.9",
+    "i18next-http-backend": "^2.1.1",
     "jwt-decode": "^3.1.2",
     "lodash-es": "^4.17.21",
-    "moment": "^2.29.1",
-    "monaco-editor": "^0.33.0",
-    "rc-field-form": "^1.24.0",
-    "react": "17.0.2",
-    "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "17.0.2",
+    "moment": "^2.29.4",
+    "monaco-editor": "^0.34.1",
+    "postcss-nesting": "^11.1.0",
+    "rc-field-form": "^1.27.3",
+    "react": "18.2.0",
+    "react-copy-to-clipboard": "^5.1.0",
+    "react-dom": "18.2.0",
     "react-helmet": "^6.1.0",
-    "react-highlight-words": "^0.17.0",
-    "react-i18next": "^11.15.7",
-    "react-markdown": "^8.0.0",
-    "react-router-dom": "6.2.2",
-    "react-simplemde-editor": "^5.0.2",
+    "react-highlight-words": "^0.20.0",
+    "react-i18next": "^12.1.4",
+    "react-markdown": "^8.0.5",
+    "react-router-dom": "6.8.0",
+    "react-simplemde-editor": "^5.2.0",
     "react-sortable-hoc": "^2.0.0",
-    "react-syntax-highlighter": "^15.4.5",
+    "react-syntax-highlighter": "^15.5.0",
     "remark-gfm": "^3.0.1",
-    "vite-plugin-imp": "^2.1.7"
+    "vite-plugin-imp": "^2.3.1"
   },
   "devDependencies": {
     "@cypress/code-coverage": "3.9.12",
-    "@nabla/vite-plugin-eslint": "1.4.0",
+    "@nabla/vite-plugin-eslint": "1.5.0",
     "@tailwindcss/forms": "0.5.0",
-    "@testing-library/cypress": "8.0.2",
-    "@testing-library/dom": "8.11.3",
-    "@testing-library/jest-dom": "5.16.2",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "7.0.2",
-    "@testing-library/user-event": "13.5.0",
+    "@testing-library/cypress": "9.0.0",
+    "@testing-library/dom": "8.20.0",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
     "@types/crypto-js": "^4.1.1",
     "@types/css-mediaquery": "0.1.1",
     "@types/lodash-es": "^4.17.6",
     "@types/react": "17.0.40",
-    "@types/react-copy-to-clipboard": "^5.0.2",
+    "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/react-dom": "17.0.13",
     "@types/react-highlight-words": "^0.16.4",
     "@types/react-router-dom": "5.3.3",
     "@types/react-syntax-highlighter": "^13.5.2",
-    "@types/testing-library__jest-dom": "^5.14.3",
+    "@types/testing-library__jest-dom": "^5.14.5",
     "@typescript-eslint/eslint-plugin": "5.14.0",
     "@typescript-eslint/parser": "5.14.0",
-    "@vitejs/plugin-react": "1.2.0",
-    "autoprefixer": "10.4.2",
+    "@vitejs/plugin-react": "3.0.1",
+    "autoprefixer": "10.4.13",
     "c8": "7.11.0",
     "commitizen": "4.2.4",
     "css-mediaquery": "0.1.2",
-    "cypress": "9.5.1",
+    "cypress": "12.4.1",
     "cz-conventional-changelog": "3.3.0",
-    "easymde": "^2.16.1",
+    "easymde": "^2.18.0",
     "eslint": "8.11.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-base": "15.0.0",
@@ -105,32 +105,30 @@
     "eslint-plugin-testing-library": "5.0.6",
     "eslint-plugin-unicorn": "41.0.0",
     "husky": "7.0.4",
-    "i18next-parser": "^6.0.0",
-    "jsdom": "19.0.0",
-    "less": "^4.1.2",
+    "i18next-parser": "^7.6.0",
+    "jsdom": "21.1.0",
     "lint-staged": "12.3.5",
     "msw": "0.39.1",
     "npm-run-all": "4.1.5",
-    "postcss": "8.4.8",
-    "postcss-less": "^6.0.0",
+    "postcss": "8.4.21",
     "prettier": "2.5.1",
-    "prettier-plugin-tailwindcss": "0.1.8",
-    "rollup": "2.70.0",
-    "rollup-plugin-istanbul": "3.0.0",
+    "prettier-plugin-tailwindcss": "0.2.2",
+    "rollup": "3.11.0",
+    "rollup-plugin-istanbul": "4.0.0",
     "start-server-and-test": "1.14.0",
     "stylelint": "14.5.3",
     "stylelint-config-prettier": "9.0.3",
     "stylelint-config-standard": "25.0.0",
-    "swagger-typescript-api": "^9.3.1",
-    "tailwindcss": "3.0.23",
-    "typescript": "4.6.2",
-    "vite": "2.8.6",
-    "vite-plugin-pwa": "0.11.13",
-    "vite-tsconfig-paths": "3.4.1",
-    "vitest": "0.6.0",
+    "swagger-typescript-api": "^12.0.2",
+    "tailwindcss": "3.2.4",
+    "typescript": "4.9.4",
+    "vite": "4.0.4",
+    "vite-plugin-pwa": "0.14.1",
+    "vite-tsconfig-paths": "4.0.5",
+    "vitest": "0.28.3",
     "whatwg-fetch": "3.6.2",
-    "workbox-build": "6.5.1",
-    "workbox-window": "6.5.1"
+    "workbox-build": "^6.5.4",
+    "workbox-window": "^6.5.4"
   },
   "browserslist": {
     "production": "Edge >= 18, Firefox >= 60, Chrome >= 61, Safari >= 11, Opera >= 48",
