@@ -6,7 +6,7 @@ import {
   TranslationOutlined,
   UserOutlined
 } from '@ant-design/icons'
-import { Dropdown, Menu, Modal, Space } from 'antd'
+import { Button, Dropdown, Modal, Space } from 'antd'
 import Gravatar from 'components/Gravatar'
 import LangSelect from 'components/LangSelect'
 import { useAccess, useAuth } from 'models'
@@ -14,6 +14,7 @@ import type React from 'react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Link, useLocation } from 'react-router-dom'
+import type { MenuItemsWithPermission, MenuItemWithPermission } from 'types'
 
 export const Index: React.FC = () => {
   const [modalVisible, setModalVisible] = useState(false)
@@ -22,39 +23,52 @@ export const Index: React.FC = () => {
   const location = useLocation()
   const access = useAccess()
 
-  const subMenu = (
-    <Menu>
-      <Menu.Item key='profile' icon={<UserOutlined />}>
+  const items: MenuItemsWithPermission = [
+    {
+      key: 'profile',
+      icon: <UserOutlined />,
+      label: (
         <Link to={`/user/${auth.user?.username ?? ''}`}>
           {t('UserMenuItem.menu.profile')}
         </Link>
-      </Menu.Item>
-      <Menu.Divider key='divider-1' />
-      <Menu.Item
-        key='switchLang'
-        icon={<TranslationOutlined />}
-        onClick={(): void => {
-          setModalVisible(true)
-        }}
-      >
-        {t('UserMenuItem.menu.switchLang')}
-      </Menu.Item>
-      <Menu.Item key='settings' icon={<SettingOutlined />}>
-        <Link to='/preference'>{t('UserMenuItem.menu.settings')}</Link>
-      </Menu.Item>
-      <Menu.Divider key='divider-2' />
-      {access.isRoot ? (
-        <>
-          <Menu.Item key='admin' icon={<ApartmentOutlined />}>
-            <Link to='/admin'>{t('UserMenuItem.menu.admin')}</Link>
-          </Menu.Item>
-          <Menu.Divider key='divider-3' />
-        </>
-      ) : null}
-      <Menu.Item key='logout' icon={<LogoutOutlined />} danger>
-        <Link to='/logout'>{t('UserMenuItem.menu.logout')}</Link>
-      </Menu.Item>
-    </Menu>
+      )
+    },
+    {
+      type: 'divider'
+    },
+    {
+      key: 'switchLang',
+      icon: <TranslationOutlined />,
+      onClick: () => setModalVisible(true),
+      label: t('UserMenuItem.menu.switchLang')
+    },
+    {
+      key: 'settings',
+      icon: <SettingOutlined />,
+      label: <Link to='/preference'>{t('UserMenuItem.menu.settings')}</Link>
+    },
+    {
+      type: 'divider'
+    },
+    {
+      key: 'admin',
+      icon: <ApartmentOutlined />,
+      label: <Link to='/admin'>{t('UserMenuItem.menu.admin')}</Link>,
+      access: access.isRoot
+    },
+    {
+      type: 'divider',
+      access: access.isRoot
+    },
+    {
+      key: 'logout',
+      icon: <LogoutOutlined />,
+      danger: true,
+      label: <Link to='/logout'>{t('UserMenuItem.menu.logout')}</Link>
+    }
+  ]
+  const filteredItems: MenuItemsWithPermission = items.filter(
+    (item: MenuItemWithPermission) => item.access !== false
   )
 
   if (auth.user) {
@@ -62,7 +76,7 @@ export const Index: React.FC = () => {
       <div>
         <Dropdown
           trigger={['click']}
-          overlay={subMenu}
+          menu={{ items: filteredItems }}
           placement='bottomRight'
           arrow
         >
@@ -73,7 +87,7 @@ export const Index: React.FC = () => {
         </Dropdown>
         <Modal
           title={t('UserMenuItem.menu.switchLang')}
-          visible={modalVisible}
+          open={modalVisible}
           onOk={(): void => {
             setModalVisible(false)
           }}
@@ -88,9 +102,11 @@ export const Index: React.FC = () => {
   }
 
   return (
-    <Link to={`/login?from=${location.pathname}`}>
-      {t('UserMenuItem.menu.login')}
-    </Link>
+    <div>
+      <Button type='primary' ghost href={`/login?from=${location.pathname}`}>
+        {t('UserMenuItem.menu.login')}
+      </Button>
+    </div>
   )
 }
 
