@@ -8,74 +8,74 @@ import { ErrorCode } from 'utils/service'
 import MainLayout from './MainLayout'
 
 const Index: React.FC = () => {
-	const { domainUrl } = useParams<{ domainUrl: string }>()
-	const { fetchDomain, errorCode, loading: domainLoading } = useDomain()
+  const { domainUrl } = useParams<{ domainUrl: string }>()
+  const { fetchDomain, errorCode, loading: domainLoading } = useDomain()
 
-	useEffect(() => {
-		fetchDomain(domainUrl)
-		return () => {
-			fetchDomain(null) // Clear the current model
-		}
-	}, [domainUrl, fetchDomain])
+  useEffect(() => {
+    fetchDomain(domainUrl)
+    return () => {
+      fetchDomain(null) // Clear the current model
+    }
+  }, [domainUrl, fetchDomain])
 
-	if (!domainLoading && errorCode) {
-		let errorTitle: string
-		let errorMessage: string
+  if (!domainLoading && errorCode) {
+    let errorTitle: string
+    let errorMessage: string
 
-		// TODO: error msg i18n & image
-		switch (errorCode) {
-			case ErrorCode.DomainNotFoundError: {
-				errorTitle = 'Domain Not Found'
-				errorMessage = 'Please check your URL.'
+    // TODO: error msg i18n & image
+    switch (errorCode) {
+      case ErrorCode.DomainNotFoundError: {
+        errorTitle = 'Domain Not Found'
+        errorMessage = 'Please check your URL.'
 
-				break
-			}
+        break
+      }
 
-			case ErrorCode.DomainUserNotFoundError: {
-				errorTitle = 'User Not Found in Domain'
-				errorMessage = 'You are not a member of this domain.'
+      case ErrorCode.DomainUserNotFoundError: {
+        errorTitle = 'User Not Found in Domain'
+        errorMessage = 'You are not a member of this domain.'
 
-				break
-			}
+        break
+      }
 
-			case ErrorCode.DomainRoleNotFoundError: {
-				errorTitle = 'Domain Role Not Found'
-				errorMessage = 'Please contact the domain administrator.'
+      case ErrorCode.DomainRoleNotFoundError: {
+        errorTitle = 'Domain Role Not Found'
+        errorMessage = 'Please contact the domain administrator.'
 
-				break
-			}
+        break
+      }
 
-			case 403: {
-				errorTitle = 'No Permission'
-				errorMessage = 'You are not a member of this domain.'
+      case 403: {
+        errorTitle = 'No Permission'
+        errorMessage = 'You are not a member of this domain.'
 
-				break
-			}
+        break
+      }
 
-			default: {
-				errorTitle = 'Unknown Error'
-				errorMessage = 'Failed to load domain info.'
-			}
-		}
+      default: {
+        errorTitle = 'Unknown Error'
+        errorMessage = 'Failed to load domain info.'
+      }
+    }
 
-		return (
-			<Result
-				className='mt-16'
-				status='404'
-				title={errorTitle}
-				subTitle={errorMessage}
-			/>
-		)
-	}
+    return (
+      <Result
+        className='mt-16'
+        status='404'
+        title={errorTitle}
+        subTitle={errorMessage}
+      />
+    )
+  }
 
-	return (
-		<>
-			<DomainHeader />
-			<MainLayout>
-				<Outlet />
-			</MainLayout>
-		</>
-	)
+  return (
+    <>
+      <DomainHeader />
+      <MainLayout>
+        <Outlet />
+      </MainLayout>
+    </>
+  )
 }
 
 export default Index
