@@ -2,13 +2,10 @@ import LoadingOrError from 'components/LoadingOrError'
 import type React from 'react'
 import { Suspense } from 'react'
 import MainContentLayout from './MainContentLayout'
-import PageHeaderLayout from './PageHeaderLayout'
 
 const Index: React.FC = ({ children }) => (
-  <MainContentLayout className='pb-16'>
-    <PageHeaderLayout>
-      <Suspense fallback={<LoadingOrError />}>{children}</Suspense>
-    </PageHeaderLayout>
+  <MainContentLayout className='pb-16 pt-6'>
+    <Suspense fallback={<LoadingOrError />}>{children}</Suspense>
   </MainContentLayout>
 )
 
