@@ -5,7 +5,7 @@ export const DOMAIN_HOST = window.location.origin
 export const LOCAL_STORAGE_SETTINGS_KEY = 'JOJ_SETTINGS'
 
 export const MAIN_CONTENT_GRID = {
-  xs: 23,
+  xs: 22,
   sm: 22,
   md: 22,
   lg: 20,
