@@ -2,12 +2,12 @@ const defaultConfig = require('tailwindcss/defaultConfig')
 const formsPlugin = require('@tailwindcss/forms')
 
 module.exports = {
-	content: ['index.html', 'src/**/*.tsx'],
-	theme: {
-		fontFamily: {
-			sans: ['Inter', ...defaultConfig.theme.fontFamily.sans]
-		}
-	},
-	experimental: { optimizeUniversalDefaults: true },
-	plugins: [formsPlugin]
+  content: ['index.html', 'src/**/*.tsx'],
+  theme: {
+    fontFamily: {
+      sans: ['Inter', ...defaultConfig.theme.fontFamily.sans]
+    }
+  },
+  experimental: { optimizeUniversalDefaults: true },
+  plugins: [formsPlugin]
 }
